transactmind/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
sentence-transformers==2.2.2
scikit-learn==1.2.2
onnx==1.14.0
onnxruntime==1.15.1
skl2onnx==1.15.0
onnxruntime-tools==1.9.0
faiss-cpu==1.7.4
shap==0.41.0
pandas==2.1.0
numpy==1.25.0
langchain==0.0.206
llama-cpp-python==0.1.57
transformers==4.34.0
streamlit==1.29.0
prometheus-client==0.17.0
locust==2.19.0
pytest==7.4.0
flake8==6.0.0
black==23.12.0
gunicorn==20.1.0
python-dotenv==1.0.0
chromadb==0.3.26
psutil==5.9.5
